refactor(filemanager): 拆出 fileManagerHelpers 纯函数子模块
- 新增 utils/：fileManagerFormat（列宽测量与格式化）、fileManagerIcons（文件类型图标）、
  fileManagerTabs（标签页比较/合并/标题与开关读取）、fileManagerItems（条目/虚拟行/排序）、
  fileManagerTransfer（分块并发与下载冲突设置）、fileManagerChmodIdentity（chmod 与属主属组解析）、
  fileManagerOpeningFiles（打开中文件全局登记）
- fileManagerHelpers.tsx 992 -> 18 行（改为桶文件，保持既有导入路径不变）
